requirements.txt
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0

# Optional: Enhanced progress display for batch execution
rich>=13.0.0
